add margins

+++ pin/Infrastructure/PinElement.cs
@@ -3,7 +3,7 @@
     public class PinElement
     {
         public string src64;
-        public int top;
-        public int left;
+        public float top;
+        public float left;
     }
 }
